+++ docling_core/transforms/chunker/doc_chunk.py
@@ -1,0 +1,92 @@
+"""Data model for chunk metadata."""
+
+import logging
+import re
+from typing import ClassVar, Final, Literal, Optional
+
+from pydantic import Field, StringConstraints, field_validator
+from typing_extensions import Annotated
+
+from docling_core.search.package import VERSION_PATTERN
+from docling_core.transforms.chunker import BaseChunk, BaseMeta
+from docling_core.types.doc.document import DocItem, DocumentOrigin
+
+_VERSION: Final = "1.0.0"
+
+_KEY_SCHEMA_NAME = "schema_name"
+_KEY_VERSION = "version"
+_KEY_DOC_ITEMS = "doc_items"
+_KEY_HEADINGS = "headings"
+_KEY_CAPTIONS = "captions"
+_KEY_ORIGIN = "origin"
+
+_logger = logging.getLogger(__name__)
+
+
+class DocMeta(BaseMeta):
+    """Data model for Hierarchical Chunker chunk metadata."""
+
+    schema_name: Literal["docling_core.transforms.chunker.DocMeta"] = Field(
+        default="docling_core.transforms.chunker.DocMeta",
+        alias=_KEY_SCHEMA_NAME,
+    )
+    version: Annotated[str, StringConstraints(pattern=VERSION_PATTERN, strict=True)] = (
+        Field(
+            default=_VERSION,
+            alias=_KEY_VERSION,
+        )
+    )
+    doc_items: list[DocItem] = Field(
+        alias=_KEY_DOC_ITEMS,
+        min_length=1,
+    )
+    headings: Optional[list[str]] = Field(
+        default=None,
+        alias=_KEY_HEADINGS,
+        min_length=1,
+    )
+    captions: Optional[list[str]] = Field(  # deprecated
+        deprecated=True,
+        default=None,
+        alias=_KEY_CAPTIONS,
+        min_length=1,
+    )
+    origin: Optional[DocumentOrigin] = Field(
+        default=None,
+        alias=_KEY_ORIGIN,
+    )
+
+    excluded_embed: ClassVar[list[str]] = [
+        _KEY_SCHEMA_NAME,
+        _KEY_VERSION,
+        _KEY_DOC_ITEMS,
+        _KEY_ORIGIN,
+    ]
+    excluded_llm: ClassVar[list[str]] = [
+        _KEY_SCHEMA_NAME,
+        _KEY_VERSION,
+        _KEY_DOC_ITEMS,
+        _KEY_ORIGIN,
+    ]
+
+    @field_validator(_KEY_VERSION)
+    @classmethod
+    def check_version_is_compatible(cls, v: str) -> str:
+        """Check if this meta item version is compatible with current version."""
+        current_match = re.match(VERSION_PATTERN, _VERSION)
+        doc_match = re.match(VERSION_PATTERN, v)
+        if (
+            doc_match is None
+            or current_match is None
+            or doc_match["major"] != current_match["major"]
+            or doc_match["minor"] > current_match["minor"]
+        ):
+            raise ValueError(f"incompatible version {v} with schema version {_VERSION}")
+        else:
+            return _VERSION
+
+
+class DocChunk(BaseChunk):
+    """Data model for document chunks."""
+
+    meta: DocMeta
